+++ src/mat/impls/elemental/matelem.cxx
@@ -94,7 +94,7 @@ static PetscErrorCode MatGetInfo_Elemental(Mat A,MatInfoType flag,MatInfo *info)
   info->block_size = 1.0;
 
   if (flag == MAT_LOCAL) {
-    info->nz_allocated   = (double)(*a->emat).AllocatedMemory(); /* locally allocated */
+    info->nz_allocated   = (*a->emat).AllocatedMemory(); /* locally allocated */
     info->nz_used        = info->nz_allocated;
   } else if (flag == MAT_GLOBAL_MAX) {
     //ierr = MPIU_Allreduce(isend,irecv,5,MPIU_REAL,MPIU_MAX,PetscObjectComm((PetscObject)matin));CHKERRQ(ierr);
@@ -102,14 +102,14 @@ static PetscErrorCode MatGetInfo_Elemental(Mat A,MatInfoType flag,MatInfo *info)
     //SETERRQ(PETSC_COMM_SELF,PETSC_ERR_SUP," MAT_GLOBAL_MAX not written yet");
   } else if (flag == MAT_GLOBAL_SUM) {
     //SETERRQ(PETSC_COMM_SELF,PETSC_ERR_SUP," MAT_GLOBAL_SUM not written yet");
-    info->nz_allocated   = (double)(*a->emat).AllocatedMemory(); /* locally allocated */
+    info->nz_allocated   = (*a->emat).AllocatedMemory(); /* locally allocated */
     info->nz_used        = info->nz_allocated; /* assume Elemental does accurate allocation */
     //ierr = MPIU_Allreduce(isend,irecv,1,MPIU_REAL,MPIU_SUM,PetscObjectComm((PetscObject)A));CHKERRQ(ierr);
     //PetscPrintf(PETSC_COMM_SELF,"    ... [%d] locally allocated %g\n",rank,info->nz_allocated);
   }
 
   info->nz_unneeded       = 0.0;
-  info->assemblies        = (double)A->num_ass;
+  info->assemblies        = A->num_ass;
   info->mallocs           = 0;
   info->memory            = ((PetscObject)A)->mem;
   info->fill_ratio_given  = 0; /* determined by Elemental */
